Split HBApplication file into two
